Move calibration algorithm to separate class
